*DaysCounter - Location, popover 위치 적절하게 위치하도록 수정, iPad Done 버튼 제거. - Location, detail 화면 진입시 annotation 표시하도록 수정.

+++ AppBox3/Apps/DaysCounter/A3DaysCounterModelManager.m
@@ -358,7 +358,7 @@ static A3DaysCounterModelManager *daysCounterModelManager = nil;
     if (alertType == AlertType_Custom) {
         //return [A3Formatter stringFromDate:date format:DaysCounterDefaultDateFormat];
         NSDateComponents *comp = [[NSCalendar currentCalendar] components:NSDayCalendarUnit fromDate:date toDate:startDate options:0];
-        return [NSString stringWithFormat:@"%d %@", comp.day, comp.day > 1 ? @"days before" : @"day before"];
+        return [NSString stringWithFormat:@"%ld %@", (long)comp.day, comp.day > 1 ? @"days before" : @"day before"];
     }
     
     return [self alertStringForType:alertType];
